+++ api/admin/controllers/category.controller.js
@@ -8,6 +8,11 @@ const jwtutil = require('../../../utils/jwt.util');
 
 exports.list = async(req, res, next) =>{
     try {
+        const errors = validationResult(req);
+        if (!errors.isEmpty()) {
+            restutil.returnValidationResponse(res, errors.array());
+            return;
+        }
        db.category.findAll({
         attributes : ['id', 'name', 'createdAt'],
         order : [[ 'id', 'ASC' ]],
@@ -19,6 +24,25 @@ exports.list = async(req, res, next) =>{
         next(error);
     }
 }
+
+exports.categoryfindbyname = async(req, res, next) =>{
+    try {
+        const errors = validationResult(req);
+        if (!errors.isEmpty()) {
+            restutil.returnValidationResponse(res, errors.array());
+            return;
+        }
+        let name = req.body.name;
+
+        db.category.findOne({
+            where : { name : name }
+        }).then((data)=>{
+            restutil.returnDataResponce(res, data);
+        })
+    } catch (error) {
+        next(error)
+    }
+} 
 
 exports.categorycreate = async (req, res, next) => {
     try {
@@ -32,11 +56,8 @@ exports.categorycreate = async (req, res, next) => {
         let name = req.body.name;
         db.category.create({
             name: name,
-            customerId : customerId
-        },{
-            where : { customerId : customerId}
-        }
-        ).then(() => {
+            
+        }).then(() => {
             restutil.returnActionSuccesResponse(res);   
         })
     }
@@ -70,13 +91,17 @@ exports.update = async (req, res, next) => {
 
 exports.delete = async (req, res, next) => {
     try {
+        const errors = validationResult(req);
+        if (!errors.isEmpty()) {
+            restutil.returnValidationResponse(res, errors.array());
+            return;
+        }
         let { id } = req.params;
         db.category.destroy({
          where : { id : id}
         }).then(() => {
             restutil.returnActionSuccesResponse(res);
         })
-
     }
     catch (error) {
         next(error);
@@ -88,12 +113,20 @@ exports.validate = (method) => {
         case 'create': {
             return [
                 body('name', 'Хоосон утга зөвшөөрөхгүй').notEmpty(),
+
             ]
         }
         case 'update': {
             return [
                 body('id', 'Хоосон утга зөвшөөрөхгүй').notEmpty(),
                 body('name', 'Хоосон утга зөвшөөрөхгүй').notEmpty(),
+            ]
+        }
+        case  'changepassword' : {
+            return [
+                body('password', 'asdasdasdas').notEmpty(),
+                body('password', 'asdasdasdas').notEmpty()
+
             ]
         }
     }
